Project's, event's & task's duration is no longer translated into secs when saved

+++ db/migrate/20110311194653_create_projects.rb
@@ -6,7 +6,7 @@ class CreateProjects < ActiveRecord::Migration
       t.date :estimated_end_date
       t.date :started_on
       t.date :ended_on
-      t.float :estimated_duration
+      t.integer :estimated_duration
       t.float :actual_duration
 
       t.timestamps
